Add link to view logs
